refactor: grid view, list view 로직 app에서 분리
